[fix] Moved classes to files
