+++ src/Device/DeviceSubTypeEnum.cs
@@ -1,0 +1,108 @@
+﻿namespace Device
+{
+    /// Подтипы устройств.
+    public enum DeviceSubType
+    {
+        NONE = -1,      ///< Подтип не определен.
+
+        //V
+        V_DO1 = 1,        ///< Клапан с одним каналом управления.
+        V_DO2,            ///< Клапан с двумя каналами управления.
+        V_DO1_DI1_FB_OFF, ///< Клапан с одним каналом управления и одной обратной связью (выключенное состояние).
+        V_DO1_DI1_FB_ON,  ///< Клапан с одним каналом управления и одной обратной связью (включенное состояние).
+        V_DO1_DI2,        ///< Клапан с одним каналом управления и двумя обратными связями.
+        V_DO2_DI2,        ///< Клапан с двумя каналами управления и двумя обратными связями.
+        V_MIXPROOF,       ///< Клапан микспруф.
+        V_AS_MIXPROOF,    ///< Клапан с двумя каналами управления и двумя обратными связями с AS интерфейсом (микспруф).
+        V_BOTTOM_MIXPROOF,///< Клапан с промывкой и двумя обратными связями (донный).
+        V_AS_DO1_DI2,     ///< Клапан с одним каналом управления и двумя обратными связями с AS интерфейсом.
+        V_DO2_DI2_BISTABLE,///< Клапан с двумя каналами управления и двумя обратными связями бистабильный.
+
+        V_IOLINK_VTUG_DO1, ///< IO-Link VTUG клапан с одним каналом управления.
+
+        V_IOLINK_VTUG_DO1_FB_OFF, ///< IO-Link VTUG клапан с одним каналом управления и одной обратной связью (выключенное состояние).
+        V_IOLINK_VTUG_DO1_FB_ON,  ///< IO-Link VTUG клапан с одним каналом управления и одной обратной связью (включенное состояние).
+        V_IOLINK_MIXPROOF, ///< Клапан микспруф с IO-Link
+        V_IOLINK_DO1_DI2,  ///< Клапан с одним каналом управления и двумя обратными связями с IO-Link интерфейсом. 
+
+        //LS
+        LS_MIN = 1,     ///< Подключение по схеме минимум.
+        LS_MAX,         ///< Подключение по схеме максимум.
+
+        LS_IOLINK_MIN,  ///< IO-Link уровень. Подключение по схеме минимум.
+        LS_IOLINK_MAX,  ///< IO-Link уровень. Подключение по схеме максимум.
+
+        LS_VIRT,        ///< Виртуальный датчик уровня.
+
+        //M,  
+        M = 1,          ///< Мотор без управления частотой вращения.
+        M_FREQ,         ///< Мотор с управлением частотой вращения.
+        M_REV,          ///< Мотор с реверсом без управления частотой вращения. Реверс включается совместно.
+        M_REV_FREQ,     ///< Мотор с реверсом с управлением частотой вращения. Реверс включается совместно.
+        M_REV_2,        ///< Мотор с реверсом без управления частотой вращения. Реверс включается отдельно.
+        M_REV_FREQ_2,   ///< Мотор с реверсом с управлением частотой вращения. Реверс включается отдельно.
+
+        /// Мотор с реверсом. Реверс включается отдельно. Отдельный сигнал
+        /// аварии.   
+        M_REV_2_ERROR,
+        /// Мотор с реверсом с управлением частотой вращения. Реверс
+        /// включается отдельно. Отдельный сигнал аварии.   
+        M_REV_FREQ_2_ERROR,
+        /// <summary>
+        /// Мотор, управляемый частотником Altivar. Связь с частотником по Ethernet.
+        /// Реверс и аварии опционально.
+        /// </summary>
+        M_ATV,
+
+        FQT = 1,        ///< Счетчик.
+        FQT_F,          ///< Счетчик с расходом.
+        FQT_F_OK,       ///< Счетчик c расходом c диагностикой.
+        FQT_VIRT,       ///Виртуальный счётчик (без привязки к модулям).
+
+        //QT
+        QT = 1,         ///< Концентратомер.
+        QT_OK,          ///< Концентратомер c диагностикой.
+        QT_IOLINK,       ///< IO-Link концентратомер.
+
+        //DO
+        DO = 1,         ///Дискретный выход с привязкой к модулям
+        DO_VIRT,        ///Виртуальный дискретный выход(без привязки к модулям)
+                        ///
+        //DI
+        DI = 1,         ///Дискретный вход с привязкой к модулям
+        DI_VIRT,        ///Виртуальный дискретный вход(без привязки к модулям)
+
+        //AO
+        AO = 1,         ///Аналоговый выход с привязкой к модулям ввода-вывода
+        AO_VIRT,        ///Виртуальный аналоговый выход (без привязки к модулям ввода-вывода)
+                        ///
+        //AI
+        AI = 1,         ///Аналоговый вход с привязкой к модулям ввода-вывода
+        AI_VIRT,        ///Виртуальный аналоговый вход (без привязки к модулям ввода-вывода)
+                        ///
+        //LT
+        LT = 1,         ///Текущий уровень без дополнительных параметров
+        LT_CYL,         ///Текущий уровень для цилиндрического танка
+        LT_CONE,        ///Текущий уровень для танка с конусом
+        LT_TRUNC,       ///Текущий уровень для танка с усеченным цилиндром
+
+        LT_IOLINK,      ///< IO-Link текущий уровень без дополнительных параметров.
+
+        LT_VIRT,        ///< Виртуальный текущий уровень.
+
+        //Y
+        Y = 1,          /// Обычный пневмоостров Festo
+        DEV_VTUG_8,     /// Festo подтип DEV_VTUG_8
+        DEV_VTUG_16,    /// Festo подтип DEV_VTUG_16
+        DEV_VTUG_24,    /// Festo подтип DEV_VTUG_24
+
+        //TE
+        TE = 1,         ///< Текущая температура.
+        TE_IOLINK,      ///< IO-Link текущая температура.
+
+        //PT
+        PT = 1,         ///< Датчик давления
+        PT_IOLINK,      ///< IO-Link датчик давления
+        DEV_SPAE        ///< IO-Link датчик давления воздуха    
+    };
+}
