+++ src/Device/DeviceTypeEnum.cs
@@ -1,0 +1,32 @@
+﻿namespace Device
+{
+    /// Типы устройств.
+    public enum DeviceType
+    {
+        NONE = -1,      ///< Тип не определен.
+
+        V = 0,   ///< Клапан. 
+        VC,      ///< Управляемый клапан. 
+        M,       ///< Двигатель.
+        LS,      ///< Уровень (есть/нет).
+        TE,      ///< Температура.        
+        FS,      ///< Расход (есть/нет).
+        GS,      ///< Датчик положения. 
+        FQT,     ///< Счетчик.        
+        LT,      ///< Уровень (значение).        
+        QT,      ///< Концентрация.
+
+        HA,      ///< Звуковая сигнализация.
+        HL,      ///< Световая сигнализация.
+        SB,      ///< Кнопка.
+        DI,      ///< Дискретный входной сигнал.
+        DO,      ///< Дискретный выходной сигнал.
+        AI,      ///< Аналоговый входной сигнал.
+        AO,      ///< Аналоговый выходной сигнал.
+        WT,      ///< Датчик веса.
+        PT,      ///< Датчик давления.
+
+        Y,       ///< Пневмоостров Festo
+        DEV_VTUG,///< Пневмоостров Festo (совместимость со старыми проектами).
+    };
+}
