+++ src/Device/IODevices/FS.cs
@@ -1,0 +1,35 @@
+﻿namespace Device
+{
+    /// <summary>
+    /// Технологическое устройство - датчик наличия потока.
+    /// Параметры:
+    /// 1. P_DT - время порогового фильтра, мсек.    
+    /// </summary>
+    public class FS : IODevice
+    {
+        public FS(string fullName, string description, int deviceNumber,
+            string objectName, int objectNumber, string articleName) : base(
+                fullName, description, deviceNumber, objectName, objectNumber)
+        {
+            dSubType = DeviceSubType.NONE;
+            dType = DeviceType.FS;
+            ArticleName = articleName;
+
+            DI.Add(new IOChannel("DI", -1, -1, -1, ""));
+
+            parameters.Add("P_DT", null);
+        }
+
+        public override string Check()
+        {
+            string res = base.Check();
+
+            if (ArticleName == "")
+            {
+                res += $"\"{name}\" - не задано изделие.\n";
+            }
+
+            return res;
+        }
+    }
+}
